manim_backend/requirements.txt
flask==3.1.0
python-dotenv==1.1.0
watchdog==6.0.0
openai==1.77.0
sqlalchemy==2.0.40
boto3==1.38.9
flask-cors==5.0.1
flask-restx==1.3.0
psycopg2-binary==2.9.10